stashbooru/requirements.txt
stashapp-tools
aiohttp
ffmpeg-python
requests
python-dotenv
